Geneve openflow support.
Matching
========
On the match side a tun_metadata is supported as a new NXM_NX_TUN_METADATA.
Multiple such nxms can appear in a flow mod in any order. tun_metadata
options are learned as flows get created. ovs-vswitch can support a max of
255 bytes of options across all flows.

For the match to work properly two tables are maintained.
1. maps (class,type) -> (len,offset)

We will not support multiple len options for same (class,type).

2. maps (offset) -> (class,type,len)

When adding flows the first table gets populated.
When expanding flows the second table is consulted to account for
holes in the tun_metadata.
When adding or looking up flows the first table is consulted.

Two new commands are added to appctl to be able to dump and flush
these tables.

  tnl/meta/flush
  tnl/meta/show

Actions
=======
A new action tun_metadata is added. Multiple of these actions can appear
in a flow mod. The actions use the same table as match.

Usage
=====
Consumers of tun_metadata api need to provide.
1. datapath
2. nlattr_to_tun_metadata
3. tun_metadata_to_nlattr

The two apis convert between datapath header format and generic tun-metadata format.