Fix building of kernel module on 3.13 and 3.16 linux kernels
While building the OVS kernel module on Ubuntu 14.04, ran into a
couple of compilation issues.
On v3.13 - duplicate export of symbol "udp_set_csum". Removed the
export since the function is marked static.
On v3.16 - definition of ovs_iptunnel_handle_offloads was getting
compiled out since the definition was wrapped in
if-kernel-version < 3.12. The function-prototype in the header file
is not conditionalized, so I'm assuming that the definition wasn't
intended to be conditionalized either.

Signed-off-by: Amit Bose <bose@noironetworks.com>